backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0
pymysql>=1.1.0
cryptography>=42.0.0
slowapi>=0.1.9
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
# passlib 1.7.x bcrypt backend, bcrypt 4.x ile uyumsuz (__about__.__version__)
passlib[bcrypt]>=1.7.4,<1.8.0
bcrypt>=3.2.0,<4.0.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth>=2.23.0
tzdata>=2024.1
